requirements.txt
blinker==1.9.0
bcrypt==4.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
face-recognition-models==0.3.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flatbuffers==23.5.26
git-filter-repo==2.47.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
py4j==0.10.9.5
PyJWT==2.10.1
pymongo==4.12.0
python-dotenv==1.1.0
qrcode==8.1
requests==2.32.3
text-unidecode==1.3
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3
twilio==9.5.2

